tipoarte/requirements.txt
#paquetes instalados para el dockerfile

tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
requests==2.21.0
requests-unixsocket==0.1.5
Pillow==6.2.1
opencv-python==4.1.2.30
pandas==0.25.2
numpy==1.17.3
matplotlib==3.1.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0





